pending objects

+++ openapi3filter/validate_request_test.go
@@ -367,7 +367,7 @@ func TestValidateQueryParams(t *testing.T) {
 			},
 			query: "param[obj]=1",
 			want: map[string]interface{}{
-				"obj": float64(1),
+				"obj": int64(1),
 			},
 		},
 		{
@@ -383,6 +383,23 @@ func TestValidateQueryParams(t *testing.T) {
 				"obj": true,
 			},
 		},
+		// FIXME: cannot build objects
+		// {
+		// 	name: "deepObject explode nested object oneOf - object",
+		// 	param: &openapi3.Parameter{
+		// 		Name: "param", In: "query", Style: "deepObject", Explode: explode,
+		// 		Schema: objectOf(
+		// 			"obj", oneofSchemaObject,
+		// 		),
+		// 	},
+		// 	query: "param[obj][id2]=1&param[obj][name2]=abc",
+		// 	want: map[string]interface{}{
+		// 		"obj": map[string]interface{}{
+		// 			"id2":   "1",
+		// 			"name2": "abc",
+		// 		},
+		// 	},
+		// },
 		//
 		//
 	}
